requirements.txt
openai>=1.0.0
wikipedia>=1.4.0
requests>=2.31.0
pydub>=0.25.1
streamlit>=1.28.0
python-dotenv>=1.0.0
pytest>=7.0.0

